Clean up parser a bit
Would like to switch over the tests for the older POC parser, but that's
going to take more time than I have most nights this week. I feel as though
it took a solid week to understand the relationship between yacc and lex and
even more time to learn, forget and relearn how yacc's grammar relates to
what it end up returning.

My immediate goal is to get some concept of ast Nodes into the yacc grammar
so I can port the tests to it.