Strip down parser
Teasing apart some of the code pulled in from the gocalc example.

What I find most annoying about goyacc is the lexer interface. There seems
to be one good way of implementing it. What I had previously was simpler,
less error prone to extend, but difficult to get working with the yacc parser.
I believe this has something to do with the rules and how digits are parsed.