Replace old lexer with hot new lexer
All in all, this was an interesting experience. I think there's a huge
opportunity to pull out the majority of the methods on the lexer into its
own separate package (that other people could use), while also improving
error reporting. The homegrown errors I've been using from debug logging
is helpful, but it's still hard to get good information when lexing
should have failed, but did not. Likewise, it's still hard to know exactly
why parsing failed.

I think removing some of the whitespace rules in the parser would be a
huge advantage, but that's a bit of work for another commit.