Interpreter can assign a value to a variable
This should work with any sort of variable, either a global variable,
regular scoped variable, instance or class variable. Since all of those
are different (and in fact, the interpreter has no concept of scopes yet),
I decided to pare this down to the smallest case that was easy to implement.