Add support for parsing "if; elsif; elsif; else"
In the process, made the decision that changing the AST representation
of a series of if, elsif, else conditions made more sense as a list
rather than a deeply nested structure. This is mostly for the convenience
of parsing, rather than for consuming this in the vm. Still unsure if this
will be the final form, but it's not too hard to change this again if need be.