Remove a hack for parsing assignment to a variable
Noticed that this was failing to assign to $0 when running rubyspecs.
After some thought, it didn't seem too helpful to define an additional
type that only the parser knows about. Perhaps if there were more specific
rules around what one can do with a variable it would be better, but for now
I think this is best.