Single quote strings are parsed without '
I'm not sure why this ever seemed like a *good* idea to me. Upon reflection,
it's really awful. Using the type system to discover which type of
string you actually are dealing with is a much more powerful technique and
it doesn't require you to slide off the quotes anytime you want the string
literal (e.g.: when you're treating it as a filename for Kernel#require).

Incidentally, this also adds the behavior of ignoring 'rubygems' for
Kernel#require.