Refactor how vm looks up the target of call expressions
This switch statement, while pretty simple, isn't going to cut it longterm.
We already have machinery to evaluate an AST node -- there's simply no reason
to reimplement this again. It gets a little crazy, and I should probably be
concerned about blowing the stack here, but this is fine for now.