Move newlines into capture list token
Working towards parsing a a call expression whose target is an expression
(possibly a grouped expression with parens with multiple statements), and
trying to refactor the capture list to be a bit more sane. It doesn't make
sense to capture expressions one by one, they need to be delimited by
something. Newlines, semicolons, etc, are what we need to delimit expressions,
otherwise madness ensues.