Support calling methods on groups
Groups of expressions with parentheses should be a valid target for a method
call. This required rethinking a bit of the highest level of the parser, and
I almost want to reproduce the same debug information that we have in the
lexer to make this easier to debug in the future. Something like easily being
able to know which rules were applied would make this so much easier to understand.