Always zero out the parser Statements in Run
It feels like this should be happening for free in the parser, but
ever since 72dc4ea this doesn't occur. It seems odd that this would have
changed significantly since the only rule removed that assigned to
parser.Statements was in parsing a single EOF token in a stream of tokens.