Working towards getting private methods on main to be callable
Right now if you just

def some_method; end

at the top level scope of a file, it gets added as a private method to Kernel.
Technically, this is correct, but those methods are not accessible since Kernel
is not included in Object, and main does not have access to its super, included
or extended methods yet.