Interpreter knows about instance methods
User defined classes!
Instance methods!
New instances!

Much excitement. Lots of hacks. In this, I managed to remove some horrible
hacks around the "current" context for defining a method in the vm. The win there
is that whether you're at the top level context or inside of a class declaration
is nice and delineated now.

On the downside, there are some hacks in place for attaching instance methods to
an instance of a class. In reality, the instance needs to look it up on its parent
class, because those can change. However, this seemed like the easiest path
forward for now.

It also occurs to me that Bad Things™ will happen if you redefine a class or
module right now.