Parse < and > as call expressions
Quite the oversight! Probably missed this because the other comparable
methods felt "more difficult" to parse, even though this is clearly
difficult because < plays double duty for inheritance.

Just discovered that this is **also** a method on Modules. You can use
this to determine if a Module is a subclass of another, or exactly equal.

http://www.ruby-doc.org/core-2.1.3/Module.html#method-i-3C