Improve handling of unless ... end expressions
There was some potential confusing as to whether the arguments to a call
expression should be parsed as a group inside the body of the unless.

Obviously that's completely wrong. The correct thing to do is to state that
there MUST be some sort of line separate at the end of the unless conditional
and before the list of statements that make up its body.