Parse array splat in call expressions
Also reverted some of the previous changes to parsing RETURN expressions inside
of methods. I took a look at MRI and it looks like there's a second pass that ensures
that "return" doesn't show up outside of methods. This seemed to be the best
direction moving forward since I was feeling like I was painting myself into
a corner with some of the recent parser changes.