Parse regex literals
Had a bit of a hard time finding a solution to this given that I had already
implemented the binary / operator. Seems like the best thing is to keep some
state of what was last parsed and only parse / as the beginning of a literal regex
when it is not likely to be the / operator.

In this case, you can only have a / operator after a target for that method call,
ie: after a number, variable, or other reference type. Most other times, it should
be expected to be a regular expression.

This doesn't work too well with newlines, but we can handle that when we get there.