Allow backslash at end of a line to continue expressions
The logic for how this works is surprisingly complicated, but I suppose I
shouldn't be surprised based on how loose ruby's grammar is. I believe
the tests for this are fairly sufficient, but it doesn't capture some of
the negative cases (e.g.: when parsing SHOULD fail).

Might be worth considering someday in the future how grubby behaves differently
(from a parsing / lexing perspective) than MRI. How much of a risk is it that
we don't fail quickly for the same errors?