Use 'go generate' to create the parser with goyacc
This is go 1.4 only, but I'm fine living at the bleeding edge of golang
for a while on this project.

It does feel odd to have some load bearing comments in the source code,
but I feel better about that than I do about language features such as
Java's annotations. They may look like comments, but I feel that is only
because 99% of the time you would want to ignore them.