Object initialization can return an error
This is akin to raising or throwing an exception, but since
we still don't have support for the BEGIN keyword in the interpreter
there isn't much cause for alarm.

The main reason I'm working on this now is because running rubyspec
is failing because of an init method referring to something that
*should* exist, but doesn't. Getting that error to bubble up was
useful, and drove out all of this.