Parse block arguments separately from other args
This should make it easier to implement blocks in the interpreter. Basically
We need an easy way to see if a block was provided and call it. Since blocks can
only ever be at the end of a list of args, it felt natural to let the interpreter
handle this, but I think encoding it in the actual type for a call expression isn't
a bad idea either.