Implement visibility for public, private methods
Didn't do protected yet because there was no need. Unsure how much
sense that makes, because it would be nice to close out the entire
"feature" of method visibility all at once, but it's also nice to
grow this as needed, by implementing the bare minimum as we go along.