Handle optional whitespace after . in call expressions
Crazy shit. Tried to handle this from the parser, but it got really angry.
Maybe it would honestly be easier to handle more optional whitespace from the
lexer, but I'm not sure, since sometimes it's an absolute must -- e.g.: when
there is an implicit block, from a method or class declaration, in particular.