Be more careful with default values in the parser
Found out that if you had a class with two methods, e.g.:

class Foo
  def something(bananas)
  end

  def something_else
  end
end

... then the parser THOUGHT that there were still some method args (bananas) to
apply to the #something_else method. This was the result of not explicitly nil'ing
out the $$ special variable in the comma_delimited_args_with_default_values rule.

Updating most of the rules to not allow empty bodies seems like a pretty safe idea
at this point.