Fix parser tests on edge of master
Apparently recently there were some changes in how casting an int to a
string works -- it really yields a string of one rune, which is
precisely what this test had relied upon. Apparently some people were
expecting it to do what fmt.Sprintf or strconv.Itoa do, which is not at
all what we want here. This test takes the actual ascii value of the
integer value from 33 to 126 and verifies we can parse them.